requirements.txt
pytest>=7.0.0
pytest-asyncio>=0.21.0
python-dotenv>=1.0.0
requests>=2.28.0
playwright>=1.40.0
aiohttp>=3.9.0
pydantic>=2.0.0
